+++ src/dimension.cpp
@@ -9,7 +9,7 @@
 
 REGISTER_CLASS(Dimension)
 {
-	register_constructor<float, float>();
+	register_constructor<int, int>();
 	register_method("get_width", &Dimension::get_width);
 	register_method("get_height", &Dimension::get_height);
 }
@@ -18,7 +18,7 @@ REGISTER_CLASS(Dimension)
 // Constructors
 //
 
-Dimension::Dimension(float width, float height)
+Dimension::Dimension(int width, int height)
 {
 	m_width = width;
 	m_height = height;
@@ -28,12 +28,12 @@ Dimension::Dimension(float width, float height)
 // Public Methods
 //
 
-float Dimension::get_width()
+int Dimension::get_width()
 {
 	return m_width;
 }
 
-float Dimension::get_height()
+int Dimension::get_height()
 {
 	return m_height;
 }
